requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
dnspython==1.16.0
idna==3.2
joblib==1.0.1
nltk==3.6.3
oauthlib==3.1.1
pymongo==3.12.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.1
regex==2021.9.24
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
textblob==0.15.3
tqdm==4.62.3
tweepy==4.4.0
urllib3==1.26.7
glom==20.11.0
python-dotenv==0.19.2
Deprecated==1.2.13